+++ CHANGELOG.md
@@ -16,7 +16,7 @@
 
 ### Added
 
-- `AdvancedCollapsible`: Added `isCollapsed` prop to be able to set it's state ([@BeirlaenAaron](https://github.com/BeirlaenAaron) in [#2511](https://github.com/teamleadercrm/ui/pull/2511))
+- `AdvancedCollapsible`: Added `defaultIsCollapsed` prop to be able to set it's state ([@BeirlaenAaron](https://github.com/BeirlaenAaron) in [#2511](https://github.com/teamleadercrm/ui/pull/2511))
 
 ## [18.4.0] - 2023-01-02